Update 0.0.0.5
Removed code relating to the tutorialChoice method, to be added as an extra once the main project is done.

Removed the workbench feature from the Room class, and deleted the RoomN11 class, to be added as an extra once the main project is done.

Added comments to the following classes:
Room01
Room02
Room03
Room04
Room05
Room06
Room07
Room08
Room09
Room010
Room
RoomChange

Overview:

Added comments to classes
Removed code relating to additions that will be completed after the main project.